Pull drake shared code out of shoe env folder
