+++ gym/envs/robot_locomotion_group/drake/shoe/manipulation_diagram.py
@@ -8,39 +8,26 @@ from tinydb.storages import MemoryStorage
 from pydrake.geometry import (
     Box,
     Cylinder,
-    DrakeVisualizer,
     GeometryFrame,
     GeometryInstance,
     MakePhongIllustrationProperties,
-    SceneGraph,
     Sphere)
-from pydrake.geometry.render import (
-    ClippingRange,
-    ColorRenderCamera,
-    DepthRange,
-    DepthRenderCamera,
-    MakeRenderEngineVtk,
-    RenderCameraCore,
-    RenderEngineVtkParams)
 from pydrake.math import RigidTransform, RollPitchYaw
 from pydrake.manipulation.schunk_wsg import SchunkWsgPositionController
 from pydrake.multibody.parsing import Parser
 from pydrake.multibody.plant import (
-    AddMultibodyPlantSceneGraph,
     CoulombFriction,
     MultibodyPlant
 )
-from pydrake.systems.sensors import CameraInfo
 from pydrake.multibody.tree import SpatialInertia, UnitInertia
 from pydrake.systems.controllers import InverseDynamicsController
-from pydrake.systems.framework import Diagram, DiagramBuilder
-from pydrake.systems.meshcat_visualizer import MeshcatVisualizer, ConnectMeshcatVisualizer
 from pydrake.systems.primitives import (Adder, Demultiplexer, MatrixGain,
                                         Multiplexer,
                                         PassThrough,
                                         StateInterpolatorWithDiscreteDerivative)
-from pydrake.systems.sensors import RgbdSensor
 
+from gym.envs.robot_locomotion_group.drake.drake_sim_diagram import \
+    DrakeSimDiagram
 from gym.envs.robot_locomotion_group.drake.shoe.vis_object import VisObject
 
 from gym.envs.robot_locomotion_group.drake.shoe.rope_utils import (
@@ -61,47 +48,19 @@ from gym.envs.robot_locomotion_group.drake.shoe.paths import (
     franka_combined_path
 )
 
-from gym.envs.robot_locomotion_group.drake.shoe.drake_utils import get_label_db
-from gym.envs.robot_locomotion_group.drake.shoe.camera_utils import camera_transform_from_dict
-from gym.envs.robot_locomotion_group.drake.shoe.transform_utils import transform_from_dict
+from gym.envs.robot_locomotion_group.drake.transform_utils import transform_from_dict
 
-class ManipulationDiagram(Diagram):
+class ManipulationDiagram(DrakeSimDiagram):
     def __init__(self, config):
-        Diagram.__init__(self)
-
-        dt = config['env']['mbp_dt']
-        builder = DiagramBuilder()
-        mbp, sg = AddMultibodyPlantSceneGraph(builder, dt)
+        DrakeSimDiagram.__init__(self, config["env"])
 
         self._config = config
-        self._mbp = mbp
-        self._sg = sg
-        self._builder = builder
-        self._rgbd_sensors = dict()
-        self._finalized = False
-
         self._model_ids = dict()
         self._control_mbp = dict()
         self._port_names = []
-        self._finalize_functions = []
         self._model_names_to_mask = list()
 
-        # add rendered
-        # Add renderer to scene
-        renderer_params = RenderEngineVtkParams()
-        self._renderer_name = "vtk_renderer"
-        self._sg.AddRenderer(
-            self._renderer_name, MakeRenderEngineVtk(renderer_params))
-
     # === Property accessors ========================================
-    @property
-    def mbp(self):
-        return self._mbp
-
-    @property
-    def sg(self):
-        return self._sg
-
     @property
     def model_ids(self):
         return self._model_ids
@@ -425,60 +384,12 @@ class ManipulationDiagram(Diagram):
                                 "middle_collision",
                                 CoulombFriction(0.0, 0.0))
             self._mbp.WeldFrames(self._mbp.world_frame(), self._mbp.GetFrameByName(f"middle_link", rope_model), RigidTransform(RollPitchYaw(roll, pitch, yaw), [(lx + rx) / 2.0, (ly + ry) / 2.0, (lz + rz) / 2.0]))
-            
+
     # === Add Cameras ===============================================
-    def add_rgbd_sensor(self, camera_name, sensor_config):
-        """
-        Adds Rgbd camera to the diagram
-        """
-        builder = self._builder
-
-        width = sensor_config['width']
-        height = sensor_config['height']
-        fov_y = sensor_config['fov_y']
-        z_near = sensor_config['z_near']
-        z_far = sensor_config['z_far']
-        renderer_name = self._renderer_name
-
-        # This is in right-down-forward convention
-        X_W_camera = camera_transform_from_dict(sensor_config)
-        color_camera = ColorRenderCamera(
-            RenderCameraCore(
-                renderer_name,
-                CameraInfo(width, height, fov_y),
-                ClippingRange(z_near, z_far),
-                RigidTransform()
-            ), False)
-        depth_camera = DepthRenderCamera(color_camera.core(),
-                                             DepthRange(z_near, z_far))
-
-
-
-        # add camera system
-        camera = builder.AddSystem(RgbdSensor(parent_id=self._sg.world_frame_id(), X_PB=X_W_camera,
-                        color_camera=color_camera,
-                        depth_camera=depth_camera))
-        builder.Connect(self._sg.get_query_output_port(),
-                        camera.query_object_input_port())
-
-        self._rgbd_sensors[camera_name] = camera
-
     def add_rgbd_sensors_from_config(self, config):
-        if not config['env']['rgbd_sensors']["enabled"]:
-            return
-        for camera_name, sensor_config in iteritems(config['env']['rgbd_sensors']['sensor_list']):
-            self.add_rgbd_sensor(camera_name, sensor_config)
+        super(ManipulationDiagram, self).add_rgbd_sensors_from_config(config["env"])
 
     # === Add visualizers ===========================================
-    def connect_to_meshcat(self):
-        meshcat = ConnectMeshcatVisualizer(self._builder, scene_graph=self._sg,
-                                           zmq_url="tcp://127.0.0.1:6000",
-                                           draw_period=1)
-        return meshcat
-
-    def connect_to_drake_visualizer(self):
-        DrakeVisualizer.AddToBuilder(builder=self._builder, scene_graph=self._sg)
-
     def add_vis_object(self, name, color):
         source_id = self._sg.RegisterSource(name)
         frame_id = self._sg.RegisterFrame(source_id, GeometryFrame(f"{name}_frame"))
@@ -494,31 +405,9 @@ class ManipulationDiagram(Diagram):
 
     # === Finalize the completed diagram ============================
     def finalize(self):
-        self._mbp.Finalize()
-        self._finalized = True
-
-        for func in self._finalize_functions:
-            func()
-
+        super(ManipulationDiagram, self).finalize()
         self._port_names.extend(["pose_bundle", "contact_results",
-                                 "plant_continuous_state", "geometry_poses", "spatial_input"])
-        self._builder.ExportOutput(self._sg.get_pose_bundle_output_port(),
-                                   "pose_bundle")
-        self._builder.ExportOutput(self._mbp.get_contact_results_output_port(),
-                                   "contact_results")
-        self._builder.ExportOutput(self._mbp.get_state_output_port(),
-                                   "plant_continuous_state")
-        self._builder.ExportOutput(self._mbp.get_geometry_poses_output_port(),
-                                   "geometry_poses")
-        self._builder.ExportInput(self._mbp.get_applied_spatial_force_input_port(),
-                                  "spatial_input")
-        self._builder.ExportOutput(self._mbp.get_body_poses_output_port(),
-                                   "body_poses")
-
-        self._builder.BuildInto(self)
-
-    def is_finalized(self):
-        return self._finalized
+                                 "plant_continuous_state", "geometry_poses", "spatial_input", "body_poses"])
 
     # === State getters & setters ===================================
     def get_model_state(self, context, model_name):
@@ -543,29 +432,6 @@ class ManipulationDiagram(Diagram):
 
     def set_pusher_state(self, context, q, v):
         return self.set_model_state(context, "pusher", q, v)
-
-    def get_image_observations_single_sensor(self, sensor_name, context):
-        assert self.is_finalized()
-        sensor = self._rgbd_sensors[sensor_name]
-        sensor_context = self.GetSubsystemContext(sensor, context)
-        rgb = sensor.color_image_output_port().Eval(sensor_context)
-        depth_32F = sensor.depth_image_32F_output_port().Eval(sensor_context)
-        depth_16U = sensor.depth_image_16U_output_port().Eval(sensor_context)
-        label = sensor.label_image_output_port().Eval(sensor_context)
-
-        return {'rgb': np.copy(rgb.data[:, :, :3]),
-                'depth_32F': np.copy(depth_32F.data).squeeze(),
-                'depth_16U': np.copy(depth_16U.data).squeeze(),
-                'label': np.copy(label.data).squeeze()}
-
-    def get_image_observations(self, context):
-        d = dict()
-        for sensor_name in self._rgbd_sensors:
-            d[sensor_name] = self.get_image_observations_single_sensor(sensor_name, context)
-        return d
-
-    def get_label_db(self):
-        return get_label_db(self._mbp)
 
     def get_labels_to_mask(self):
         """
